--- app.py
@@ -1,4 +1,0 @@
-from flask import Flask
-from app import create_app
-
-app: Flask = create_app()
